refactor: factor out AstParser from DocParser
